zoom a little less on startup

+++ src/iter_algos.c
@@ -6,7 +6,7 @@
 /*   By: mjacq <mjacq@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/06 14:28:27 by mjacq             #+#    #+#             */
-/*   Updated: 2021/12/06 18:39:35 by mjacq            ###   ########.fr       */
+/*   Updated: 2021/12/06 19:30:36 by mjacq            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,7 +44,7 @@ int	iter_mandelbrot(t_coordinates *c, int max_iter)
 {
 	t_coordinates	z0;
 
-	if (powf(c->x + 1, 2) + powf(c->y, 2) < 1.0 / 16)
+	if (powf(c->x + 1, 2) + powf(c->y, 2) < (float)1.0 / 16)
 		return (max_iter);
 	z0 = (t_coordinates){};
 	return (iter_julia(c, &z0, max_iter));
